requirements.txt
elasticsearch
gfal2-python
rucio-clients
slackclient
python-crontab
pyyaml
